requirements.txt
langchain
langchain-community
langchain-ollama
neo4j
streamlit
tavily-python
python-dotenv